Don't inject min/max exprs in the bounds of a promise_clamped of a constant
No effect in public Halide, but I hear it causes at least one regression
inside Google. The reason for injecting the min/max exprs in the case
where the first arg doesn't vary has since gone away now that we push
producers inside the guard-with-if if statement of a consumer.